+++ Runtime/StateGraph/States/BaseSubStateMachineState.cs
@@ -22,20 +22,20 @@ namespace Nonatomic.VSM2.StateGraph.States
 		private bool _started;
 		private bool _entered;
 
-		public override void OnAwakeState()
+		public override void OnAwake()
 		{
 			CreateStateMachine();
 			ReplaceModelWithActiveModel();
 		}
 		
-		public override void OnStartState()
+		public override void OnStart()
 		{
 			SubStateMachine?.Start();
 			_started = true;
 		}
 
 		[Enter]
-		public override void OnEnterState()
+		public override void OnEnter()
 		{
 			if(SubStateMachine == null) return;
 
@@ -45,17 +45,17 @@ namespace Nonatomic.VSM2.StateGraph.States
 			SubStateMachine.Enter();
 		}
 
-		public override void OnUpdateState()
+		public override void OnUpdate()
 		{
 			SubStateMachine?.Update();
 		}
 
-		public override void OnFixedUpdateState()
+		public override void OnFixedUpdate()
 		{
 			SubStateMachine?.FixedUpdate();
 		}
 
-		public override void OnExitState()
+		public override void OnExit()
 		{
 			if(SubStateMachine == null) return;
 			
@@ -64,7 +64,7 @@ namespace Nonatomic.VSM2.StateGraph.States
 			_entered = false;
 		}
 
-		public override void OnDestroyState()
+		public override void OnDestroy()
 		{
 			ReplaceModelWithOriginalModel();