+++ Runtime/StateGraph/States/DelayState.cs
@@ -15,12 +15,12 @@ namespace Nonatomic.VSM2.StateGraph.States
 		private float _elapsedTime;
 
 		[Enter]
-		public override void OnEnterState()
+		public override void OnEnter()
 		{
 			_elapsedTime = 0f;
 		}
 		
-		public override void OnUpdateState()
+		public override void OnUpdate()
 		{
 			_elapsedTime += Time.deltaTime;
 
@@ -28,7 +28,7 @@ namespace Nonatomic.VSM2.StateGraph.States
 			OnComplete?.Invoke();
 		}
 
-		public override void OnExitState()
+		public override void OnExit()
 		{
 			//...
 		}