+++ Runtime/StateGraph/States/EntryState.cs
@@ -13,12 +13,12 @@ namespace Nonatomic.VSM2.StateGraph.States
 		public event Action OnEntry;
 
 		[Enter]
-		public override void OnEnterState()
+		public override void OnEnter()
 		{
 			OnEntry?.Invoke();
 		}
 
-		public override void OnExitState()
+		public override void OnExit()
 		{
 			//..
 		}