+++ Runtime/StateGraph/States/RandomThreeState.cs
@@ -14,7 +14,7 @@ namespace Nonatomic.VSM2.StateGraph.States
 		private Random _random = new();
 
 		[Enter]
-		public override void OnEnterState()
+		public override void OnEnter()
 		{
 			switch (_random.Next(0, 2))
 			{
@@ -30,7 +30,7 @@ namespace Nonatomic.VSM2.StateGraph.States
 			}
 		}
 
-		public override void OnExitState()
+		public override void OnExit()
 		{
 			//...
 		}