+++ Runtime/StateGraph/States/RandomTwoState.cs
@@ -13,7 +13,7 @@ namespace Nonatomic.VSM2.StateGraph.States
 		private Random _random = new();
 
 		[Enter]
-		public override void OnEnterState()
+		public override void OnEnter()
 		{
 			switch (_random.Next(0, 1))
 			{
@@ -26,7 +26,7 @@ namespace Nonatomic.VSM2.StateGraph.States
 			}
 		}
 
-		public override void OnExitState()
+		public override void OnExit()
 		{
 			//...
 		}